consumer/requirements.txt
kafka-python
python-dotenv==1.0.1
cassandra-driver